Merge

+++ src/networkmessage.cpp
@@ -1,6 +1,8 @@
 /**
+ * @file networkmessage.cpp
+ * 
  * The Forgotten Server - a free and open-source MMORPG server emulator
- * Copyright (C) 2019 Mark Samman <mark.samman@gmail.com>
+ * Copyright (C) 2020 Mark Samman <mark.samman@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -26,7 +28,7 @@
 
 int32_t NetworkMessage::decodeHeader()
 {
-	int32_t newSize = static_cast<int32_t>(buffer[0] | buffer[1] << 8);
+	int32_t newSize = buffer[0] | buffer[1] << 8;
 	info.length = newSize;
 	return info.length;
 }
